fix : UserInfo & Wish

+++ backend/service/user-service/src/main/java/shop/biday/service/impl/AddressServiceImpl.java
@@ -22,17 +22,6 @@ public class AddressServiceImpl implements AddressService {
     private final UserInfoUtils userInfoUtils;
 
     @Override
-    public Flux<AddressDocument> findAll() {
-        return addressRepository.findAll();
-    }
-
-    @Override
-    public Mono<AddressDocument> findById(String id) {
-        return addressRepository.findById(id)
-                .switchIfEmpty(Mono.error(new RuntimeException("주소를 찾지 못했습니다.")));
-    }
-
-    @Override
     public Mono<AddressDocument> save(String userId, AddressModel addressModel) {
 
         AddressType addressType;
@@ -65,17 +54,13 @@ public class AddressServiceImpl implements AddressService {
     }
 
     @Override
-    public Mono<Boolean> existsById(String id) {
-        return addressRepository.existsById(id);
-    }
+    public Mono<Boolean> deleteById(String userInfoHeader, String id) {
+        UserInfoModel userInfo = userInfoUtils.extractUserInfo(userInfoHeader);
 
-    @Override
-    public Mono<Long> count() {
-        return addressRepository.count();
-    }
+        if (userInfo.getUserId() == null) {
+            return Mono.error(new IllegalArgumentException("유저 ID는 없습니다."));
+        }
 
-    @Override
-    public Mono<Boolean> deleteById(String id) {
         return addressRepository.deleteById(id).hasElement();
     }
 
@@ -111,4 +96,25 @@ public class AddressServiceImpl implements AddressService {
         UserInfoModel userInfoModel = userInfoUtils.extractUserInfo(userInfoHeader);
         return addressRepository.findAllByUserId(userInfoModel.getUserId());
     }
+
+//    @Override
+//    public Flux<AddressDocument> findAll() {
+//        return addressRepository.findAll();
+//    }
+//
+//    @Override
+//    public Mono<AddressDocument> findById(String id) {
+//        return addressRepository.findById(id)
+//                .switchIfEmpty(Mono.error(new RuntimeException("주소를 찾지 못했습니다.")));
+//    }
+//
+//    @Override
+//    public Mono<Boolean> existsById(String id) {
+//        return addressRepository.existsById(id);
+//    }
+//
+//    @Override
+//    public Mono<Long> count() {
+//        return addressRepository.count();
+//    }
 }
